# Improve performance of vcf-variant->protein-hgvs