--- a/project.clj
+++ b/project.clj
@@ -7,7 +7,7 @@
   :dependencies [[org.clojure/clojure "1.9.0" :scope "provided"]
                  [org.clojure/tools.logging "0.4.1"]
                  [clj-hgvs "0.2.4"]
-                 [cljam "0.6.0"]
+                 [cljam "0.7.0"]
                  [org.apache.commons/commons-compress "1.18"]
                  [proton "0.1.6"]]
   :plugins [[lein-cloverage "1.0.13"]