Improves performance of vcf-variant->protein-hgvs by optimizing amino acid conversion and sequence retrieving. That can be checked with libra benchmark code.

before (commit 3ec5ed6):

```console
$ lein libra :only varity.vcf-to-hgvs-bench

Measuring varity.vcf-to-hgvs-bench

vcf-variant->protein-hgvs-bench (vcf_to_hgvs_bench.clj:19)

  time: 97.077914 ms, sd: 113.403155 µs
```

after (commit ac184f7):

```console
$ lein libra :only varity.vcf-to-hgvs-bench

Measuring varity.vcf-to-hgvs-bench

vcf-variant->protein-hgvs-bench (vcf_to_hgvs_bench.clj:19)

  time: 26.393281 ms, sd: 597.575155 ns
```

I confirmed `lein test :all` passed.